Pull out general-purpose syncing methods that are independent from Iroh
